refactor(bench): enhance benchmarking structure and update scripts
- Updated the `AGENTS.md` file to improve clarity on directory purposes and added new sections for atomic benchmarks and exploratory scripts.
- Introduced a new `bench/bench-atomic.ts` script for dispatching atomic micro-benchmarks, consolidating benchmark execution.
- Removed outdated benchmark scripts and reorganized existing ones for better maintainability.
- Updated `package.json` to reflect changes in benchmark commands and improved script execution consistency.
- Enhanced performance measurement documentation and clarified usage instructions across various benchmark scripts.
